pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "activation-as-view"
version = "0.1.0"
description = "Activation-as-View research monorepo for LeWM x Qwen3-72B"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "pillow>=10.0",
  "h5py>=3.11",
  "pyyaml>=6.0",
  "matplotlib>=3.8",
  "typer>=0.12",
  "tensorboard>=2.16",
]

[project.optional-dependencies]
torch = ["torch>=2.2", "transformers>=4.45", "accelerate>=0.33"]
dev = ["pytest>=8.2"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
